+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Changes
-+ **8.2.5** - Update to FatDynet 0.3.0 in serial mode 
++ **8.2.5** - Added CluCoreProcessor, which uses all our tools + Stanford's numeric entity recognizer. 
++ **8.2.5** - Update to FatDynet 0.3.0 in serial mode. This fixes memory leaks and crashed in the DyNet Scala wrapper. 
 + **8.2.4** - Bug fix: the statement below holds even when the dependency graph is deserialized via the DocumentSerializer.
 + **8.2.4** - Bug fix: we now guarantee that the Stanford dependency graph produced by CoreNLP has the same number of nodes as the words in the sentence.
 + **8.2.3** - Do not allow loops in enhanced semantic roles.